Add mod.bundle tag to the TagManager
I took a look at a newly copied ktanemodkit to test wawa stuff and noticed that I had the mod.bundle tag available in my project despite not having any bundles labelled with the tag. After looking on google, it appears the tag is saved in ProjectSettings/TagManager.asset, and I can confirm that the tag exists in some of my projects and it showed up in my test project, because I used the ktanemodkit to build mod.bundle before after deleting it.